Don't save the exceptiontype at layerlevel, but at geoservice level
