+++ viewer-admin/src/main/webapp/META-INF/context.xml
@@ -2,6 +2,8 @@
 <!-- crossContext="true" is required to enable the viewer-admin to load the 
   components.json file and component config javascript sources from the viewer -->
 <Context antiJARLocking="true" crossContext="true" disableURLRewriting="true" path="/viewer-admin">
+     <Parameter name="componentregistry.path" override="false" value="/viewer-html/components,/opt/flamingo_data/flamingo_3rdparty/dbk-component,/opt/flamingo_data/flamingo_3rdparty/ro-tercera-component"/>
+
   <!-- Customized context parameters; see web.xml -->
   <!--Parameter name="viewer.url" value="/viewer" override="false"/-->
   <!--Parameter name="monitoring.mail.from.email" value="no-reply@b3partners.nl" override="false"/-->
